nodejs/00-node-fundamentals/01-file-system-operations/src/index.ts
import fs from 'fs/promises';
import fsSync from 'fs';
import path from 'path';

export type FileOptions = {
  encoding?: BufferEncoding;
  mode?: number;
  flag?: string;
};

export class FileManager {
  private basePath: string;

  constructor(basePath: string) {
    this.basePath = basePath;
  }

  /**
   * Read file contents
   * @param filePath - Path to file
   * @param options - File options
   */
  async readFile(filePath: string, options?: FileOptions): Promise<string | Buffer> {
    throw new Error('Not implemented');
  }

  /**
   * Write data to file
   * @param filePath - Path to file
   * @param data - Data to write
   * @param options - File options
   */
  async writeFile(filePath: string, data: string | Buffer, options?: FileOptions): Promise<void> {
    throw new Error('Not implemented');
  }

  /**
   * Check if file exists
   * @param filePath - Path to file
   */
  async exists(filePath: string): Promise<boolean> {
    throw new Error('Not implemented');
  }

  /**
   * Watch file for changes
   * @param filePath - Path to file
   * @param onChange - Change callback
   */
  watch(filePath: string, onChange: (eventType: string, filename: string) => void): void {
    throw new Error('Not implemented');
  }

  /**
   * List directory contents
   * @param dirPath - Path to directory
   */
  async readDir(dirPath: string): Promise<string[]> {
    throw new Error('Not implemented');
  }

  /**
   * Helper to resolve paths relative to base path
   * @private
   */
  private resolvePath(filePath: string): string {
    return path.join(this.basePath, filePath);
  }
}
